Abstract performance testing workflow into a bash script.
Use perf/run-benchmarks to run a benchmark comparison between
the last version of the current branch and a previous version
of the current branch, determined by how many commits back to
go (passed as an argument to the script).

Don't forget to chmod +x the script!

The script can also be run as' npm run perf' or, if you want
to pass arguments to it, 'npm run perf -- <arg>'.