chore(ramda): Remove dependency on external library
*TLDR* Reduce final module size and dependency clutter.

Ramda is cool. Using it can help make code more functional and
expressive. However, it’s not always a good idea to couple a library
to a third party dependency if we can solve the problem with a few
lines of code.

I guess this wouldn't be so much of an issue if the build output
something which did not contain the `require` for the ES module.
The main issue with requiring this third party dependency is that
it increases the amount of code that needs to be downloaded, since
we have another library which is required from this one. This is
mainly a problem for those of us who are using native ES modules
rather than using bundlers with tree shaking I suppose.